Refactor providers and screens: streamline loading states, update base URLs, and enhance logout dialog

+++ lib/widgets/resgister_button.dart
@@ -27,7 +27,7 @@ class RegisterButton extends StatelessWidget {
               ),
               onPressed: () {
                 authProvider.register(
-                    name.text, phone.text as int, password.text);
+                    name.text, phone.text , password.text);
               }),
         ),
       ],
